added custom hook to useForm.js and implemented hook in CheckoutForm.js
